ugly but working
