+++ src/bulkImportWrappers.ts
@@ -14,6 +14,7 @@ export async function executeBulkImport(
   } else {
     ({ output: bulkDataResponses } = await retrieveAllBulkData(exportUrl));
   }
+  console.log(bulkDataResponses);
   if (bulkDataResponses) {
     const db = await populateDB(bulkDataResponses, location);
     const tbArr = await assembleTransactionBundles(db);
@@ -21,3 +22,7 @@ export async function executeBulkImport(
     return tbArr;
   }
 }
+
+executeBulkImport('http://localhost:3000', ':memory:')
+  .then(res => console.log(res))
+  .catch(e => console.log(e.message));
